UniqueID for SocketMessage
When I send a message from the server, I do not know what ID it will end up having on Discord.

Instead, I create an ID myself and place it in an embed footer text field so it doesn't get lost and is not as annoying.

Of course if I do not plan to use the sent message, the ID is not needed and I simply don't pass the attachID variable.